Fix cli bug where spinner overwrites prompt
When running `skpm publish <version>` for the first time, and plugin is not on directory yet, it seems to be checking infinitely. This just a cosmetic issue, however: the spinner overwrites and hides the prompt in terminal.

This is fixed by moving the creation of the spinner instance into the `print()` function, allowing you to stop the spinner at any given point, clearing it, or even declaring it null. The `print()` function will always check if the `spinner` variable is truthy, and declare `new (ora)` with default settings and text passed into print and restart the spinner.